Estudo_Django/Projeto_Django/my_api/serializer.py
from rest_framework import serializers
from my_api.models import Cadastro 

class CadastroSerializer(serializers.ModelSerializer):
  class Meta:
    model = Cadastro
    fields = ['id', 'name', 'idade', 'cpf']