Estudo_Django/Projeto_Django/requirement.txt
algoliasearch==2.6.3
algoliasearch-django==2.0.0
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
djangorestframework==3.14.0
idna==3.4
pycodestyle==2.10.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.15